Move types into base folder
